chore: Move rename doctype method to dedicated TestCase
