+++ frappe/tests/test_document.py
@@ -249,82 +249,6 @@ class TestDocument(unittest.TestCase):
 
 			self.assertEqual(cint(old_current) - 1, new_current)
 
-	def test_rename_doc(self):
-		from random import choice, sample
-
-		available_documents = []
-		doctype = "ToDo"
-
-		# data generation: 4 todo documents
-		for num in range(1, 5):
-			doc = frappe.get_doc({
-				"doctype": doctype,
-				"date": add_to_date(now(), days=num),
-				"description": "this is todo #{}".format(num)
-			}).insert()
-			available_documents.append(doc.name)
-
-		# test 1: document renaming
-		old_name = choice(available_documents)
-		new_name = old_name + '.new'
-		self.assertEqual(new_name, frappe.rename_doc(doctype, old_name, new_name, force=True))
-		available_documents.remove(old_name)
-		available_documents.append(new_name)
-
-		# test 2: merge documents
-		first_todo, second_todo = sample(available_documents, 2)
-
-		second_todo_doc = frappe.get_doc(doctype, second_todo)
-		second_todo_doc.priority = "High"
-		second_todo_doc.save()
-
-		merged_todo = frappe.rename_doc(doctype, first_todo, second_todo, merge=True, force=True)
-		merged_todo_doc = frappe.get_doc(doctype, merged_todo)
-		available_documents.remove(first_todo)
-
-		with self.assertRaises(DoesNotExistError):
-			frappe.get_doc(doctype, first_todo)
-
-		self.assertEqual(merged_todo_doc.priority, second_todo_doc.priority)
-
-		for docname in available_documents:
-			frappe.delete_doc(doctype, docname)
-
-	def test_rename_doctype(self):
-		from frappe.core.doctype.doctype.test_doctype import new_doctype
-
-		fields =[{
-			"label": "Linked To",
-			"fieldname": "linked_to_doctype",
-			"fieldtype": "Link",
-			"options": "DocType",
-			"unique": 0
-		}]
-		if not frappe.db.exists("DocType", "Rename This"):
-			new_doctype("Rename This", unique=0, fields=fields).insert()
-
-		to_rename_record = frappe.get_doc({
-			"doctype": "Rename This",
-			"linked_to_doctype": "Rename This"
-		})
-		to_rename_record.insert()
-
-		# Rename doctype
-		self.assertEqual("Renamed Doc", frappe.rename_doc("DocType", "Rename This", "Renamed Doc", force=True))
-
-		# Test if Doctype value has changed in Link field
-		renamed_doctype_record = frappe.get_doc("Renamed Doc", to_rename_record.name)
-		self.assertEqual(renamed_doctype_record.linked_to_doctype, "Renamed Doc")
-
-		# Test if there are conflicts between a record and a DocType
-		# having the same name
-		old_name = to_rename_record.name
-		new_name = "ToDo"
-		self.assertEqual(new_name, frappe.rename_doc("Renamed Doc", old_name, new_name, force=True))
-
-		frappe.delete_doc_if_exists("Renamed Doc", "ToDo")
-		frappe.delete_doc_if_exists("DocType", "Renamed Doc")
-
 	def test_non_negative_check(self):
 		frappe.delete_doc_if_exists("Currency", "Frappe Coin", 1)
